Wtf was the last commit
Last commit broke everything, this should fix it

+++ FactorioModSwapper/v15confirmation.vb
@@ -9,17 +9,21 @@ Public Class v15confirmation
     End Sub
 
     Public Sub v15continue_Click(sender As Object, e As EventArgs) Handles v15continue.Click
-        Directory.Delete(FactorioModSwitcher.modPath, True)
-        My.Computer.FileSystem.CopyDirectory(FactorioModSwitcher.v16backupVar, FactorioModSwitcher.modPath, True)
-        My.Computer.Audio.PlaySystemSound(Media.SystemSounds.Exclamation)
-        Me.Hide()
+        If Directory.Exists(FactorioModSwitcher.modPath) Then
+            Directory.Delete(FactorioModSwitcher.modPath, True)
+            Directory.CreateDirectory(FactorioModSwitcher.modPath)
+            My.Computer.FileSystem.CopyDirectory(FactorioModSwitcher.v15backupVar, FactorioModSwitcher.modPath, True)
+            My.Computer.Audio.PlaySystemSound(Media.SystemSounds.Exclamation)
+            Me.Hide()
+        Else
+            Directory.CreateDirectory(FactorioModSwitcher.modPath)
+            My.Computer.FileSystem.CopyDirectory(FactorioModSwitcher.v15backupVar, FactorioModSwitcher.modPath, True)
+            My.Computer.Audio.PlaySystemSound(Media.SystemSounds.Exclamation)
+            Me.Hide()
+        End If
     End Sub
 
     Private Sub v15confirmation_Load(sender As Object, e As EventArgs) Handles MyBase.Load
-
-    End Sub
-
-    Private Sub Label1_Click(sender As Object, e As EventArgs) Handles Label1.Click
 
     End Sub
 End Class
